+++ src/tests/cases/ui.rs
@@ -103,10 +103,38 @@ fn basic_only_processes() {
         interface: Some(String::from("interface_name")),
         raw: false,
         no_resolve: false,
+        show_dns: false,
         render_opts: RenderOpts {
             addresses: false,
             connections: false,
             processes: true,
+            total_utilization: false,
+        },
+    };
+
+    start(backend, os_input, opts);
+    let terminal_draw_events_mirror = terminal_draw_events.lock().unwrap();
+    assert_snapshot!(&terminal_draw_events_mirror[0]);
+}
+
+#[test]
+fn basic_processes_with_dns_queries() {
+    let network_frames = vec![NetworkFrames::new(vec![
+        None, // sleep
+    ]) as Box<dyn DataLinkReceiver>];
+
+    let (_, terminal_draw_events, backend) = test_backend_factory(190, 50);
+    let os_input = os_input_output(network_frames, 1);
+    let opts = Opt {
+        interface: Some(String::from("interface_name")),
+        raw: false,
+        no_resolve: false,
+        show_dns: true,
+        render_opts: RenderOpts {
+            addresses: false,
+            connections: false,
+            processes: true,
+            total_utilization: false,
         },
     };
 
@@ -127,10 +155,12 @@ fn basic_only_connections() {
         interface: Some(String::from("interface_name")),
         raw: false,
         no_resolve: false,
+        show_dns: false,
         render_opts: RenderOpts {
             addresses: false,
             connections: true,
             processes: false,
+            total_utilization: false,
         },
     };
 
@@ -151,10 +181,12 @@ fn basic_only_addresses() {
         interface: Some(String::from("interface_name")),
         raw: false,
         no_resolve: false,
+        show_dns: false,
         render_opts: RenderOpts {
             addresses: true,
             connections: false,
             processes: false,
+            total_utilization: false,
         },
     };
 
@@ -173,10 +205,12 @@ fn two_packets_only_processes() {
         interface: Some(String::from("interface_name")),
         raw: false,
         no_resolve: false,
+        show_dns: false,
         render_opts: RenderOpts {
             addresses: false,
             connections: false,
             processes: true,
+            total_utilization: false,
         },
     };
 
@@ -196,10 +230,12 @@ fn two_packets_only_connections() {
         interface: Some(String::from("interface_name")),
         raw: false,
         no_resolve: false,
+        show_dns: false,
         render_opts: RenderOpts {
             addresses: false,
             connections: true,
             processes: false,
+            total_utilization: false,
         },
     };
 
@@ -219,10 +255,12 @@ fn two_packets_only_addresses() {
         interface: Some(String::from("interface_name")),
         raw: false,
         no_resolve: false,
+        show_dns: false,
         render_opts: RenderOpts {
             addresses: true,
             connections: false,
             processes: false,
+            total_utilization: false,
         },
     };
 
@@ -244,10 +282,12 @@ fn two_windows_split_horizontally() {
         interface: Some(String::from("interface_name")),
         raw: false,
         no_resolve: false,
+        show_dns: false,
         render_opts: RenderOpts {
             addresses: true,
             connections: true,
             processes: false,
+            total_utilization: false,
         },
     };
 
@@ -268,10 +308,12 @@ fn two_windows_split_vertically() {
         interface: Some(String::from("interface_name")),
         raw: false,
         no_resolve: false,
+        show_dns: false,
         render_opts: RenderOpts {
             addresses: true,
             connections: true,
             processes: false,
+            total_utilization: false,
         },
     };
 
@@ -460,7 +502,7 @@ fn multiple_processes_with_multiple_connections() {
             "10.0.0.2",
             1337,
             4435,
-            b"Awesome, I'm from 3.3.3.3",
+            b"Greetings traveller, I'm from 3.3.3.3",
         )),
         Some(build_tcp_packet(
             "2.2.2.2",
@@ -576,6 +618,47 @@ fn sustained_traffic_from_one_process() {
 }
 
 #[test]
+fn sustained_traffic_from_one_process_total() {
+    let network_frames = vec![NetworkFrames::new(vec![
+        Some(build_tcp_packet(
+            "1.1.1.1",
+            "10.0.0.2",
+            12345,
+            443,
+            b"I have come from 1.1.1.1",
+        )),
+        None, // sleep
+        Some(build_tcp_packet(
+            "1.1.1.1",
+            "10.0.0.2",
+            12345,
+            443,
+            b"Same here, but one second later",
+        )),
+    ]) as Box<dyn DataLinkReceiver>];
+
+    let (terminal_events, terminal_draw_events, backend) = test_backend_factory(190, 50);
+
+    let os_input = os_input_output(network_frames, 3);
+    let mut opts = opts_ui();
+    opts.render_opts.total_utilization = true;
+    start(backend, os_input, opts);
+    let terminal_draw_events_mirror = terminal_draw_events.lock().unwrap();
+
+    let expected_terminal_events = vec![
+        Clear, HideCursor, Draw, Flush, Draw, Flush, Draw, Flush, Clear, ShowCursor,
+    ];
+    assert_eq!(
+        &terminal_events.lock().unwrap()[..],
+        &expected_terminal_events[..]
+    );
+
+    assert_eq!(terminal_draw_events_mirror.len(), 3);
+    assert_snapshot!(&terminal_draw_events_mirror[1]);
+    assert_snapshot!(&terminal_draw_events_mirror[2]);
+}
+
+#[test]
 fn sustained_traffic_from_multiple_processes() {
     let network_frames = vec![NetworkFrames::new(vec![
         Some(build_tcp_packet(
@@ -613,6 +696,61 @@ fn sustained_traffic_from_multiple_processes() {
 
     let os_input = os_input_output(network_frames, 3);
     let opts = opts_ui();
+    start(backend, os_input, opts);
+    let terminal_draw_events_mirror = terminal_draw_events.lock().unwrap();
+
+    let expected_terminal_events = vec![
+        Clear, HideCursor, Draw, Flush, Draw, Flush, Draw, Flush, Clear, ShowCursor,
+    ];
+    assert_eq!(
+        &terminal_events.lock().unwrap()[..],
+        &expected_terminal_events[..]
+    );
+
+    assert_eq!(terminal_draw_events_mirror.len(), 3);
+    assert_snapshot!(&terminal_draw_events_mirror[1]);
+    assert_snapshot!(&terminal_draw_events_mirror[2]);
+}
+
+#[test]
+fn sustained_traffic_from_multiple_processes_total() {
+    let network_frames = vec![NetworkFrames::new(vec![
+        Some(build_tcp_packet(
+            "1.1.1.1",
+            "10.0.0.2",
+            12345,
+            443,
+            b"I have come from 1.1.1.1",
+        )),
+        Some(build_tcp_packet(
+            "3.3.3.3",
+            "10.0.0.2",
+            1337,
+            4435,
+            b"I come from 3.3.3.3",
+        )),
+        None, // sleep
+        Some(build_tcp_packet(
+            "1.1.1.1",
+            "10.0.0.2",
+            12345,
+            443,
+            b"I have come from 1.1.1.1 one second later",
+        )),
+        Some(build_tcp_packet(
+            "3.3.3.3",
+            "10.0.0.2",
+            1337,
+            4435,
+            b"I come 3.3.3.3 one second later",
+        )),
+    ]) as Box<dyn DataLinkReceiver>];
+
+    let (terminal_events, terminal_draw_events, backend) = test_backend_factory(190, 50);
+
+    let os_input = os_input_output(network_frames, 3);
+    let mut opts = opts_ui();
+    opts.render_opts.total_utilization = true;
     start(backend, os_input, opts);
     let terminal_draw_events_mirror = terminal_draw_events.lock().unwrap();
 
@@ -695,6 +833,89 @@ fn sustained_traffic_from_multiple_processes_bi_directional() {
 
     let os_input = os_input_output(network_frames, 3);
     let opts = opts_ui();
+    start(backend, os_input, opts);
+    let terminal_draw_events_mirror = terminal_draw_events.lock().unwrap();
+
+    let expected_terminal_events = vec![
+        Clear, HideCursor, Draw, Flush, Draw, Flush, Draw, Flush, Clear, ShowCursor,
+    ];
+    assert_eq!(
+        &terminal_events.lock().unwrap()[..],
+        &expected_terminal_events[..]
+    );
+
+    assert_eq!(terminal_draw_events_mirror.len(), 3);
+    assert_snapshot!(&terminal_draw_events_mirror[1]);
+    assert_snapshot!(&terminal_draw_events_mirror[2]);
+}
+
+#[test]
+fn sustained_traffic_from_multiple_processes_bi_directional_total() {
+    let network_frames = vec![NetworkFrames::new(vec![
+        Some(build_tcp_packet(
+            "10.0.0.2",
+            "3.3.3.3",
+            4435,
+            1337,
+            b"omw to 3.3.3.3",
+        )),
+        Some(build_tcp_packet(
+            "3.3.3.3",
+            "10.0.0.2",
+            1337,
+            4435,
+            b"I was just there!",
+        )),
+        Some(build_tcp_packet(
+            "1.1.1.1",
+            "10.0.0.2",
+            12345,
+            443,
+            b"Is it nice there? I think 1.1.1.1 is dull",
+        )),
+        Some(build_tcp_packet(
+            "10.0.0.2",
+            "1.1.1.1",
+            443,
+            12345,
+            b"Well, I heard 1.1.1.1 is all the rage",
+        )),
+        None, // sleep
+        Some(build_tcp_packet(
+            "10.0.0.2",
+            "3.3.3.3",
+            4435,
+            1337,
+            b"Wait for me!",
+        )),
+        Some(build_tcp_packet(
+            "3.3.3.3",
+            "10.0.0.2",
+            1337,
+            4435,
+            b"They're waiting for you...",
+        )),
+        Some(build_tcp_packet(
+            "1.1.1.1",
+            "10.0.0.2",
+            12345,
+            443,
+            b"1.1.1.1 forever!",
+        )),
+        Some(build_tcp_packet(
+            "10.0.0.2",
+            "1.1.1.1",
+            443,
+            12345,
+            b"10.0.0.2 forever!",
+        )),
+    ]) as Box<dyn DataLinkReceiver>];
+
+    let (terminal_events, terminal_draw_events, backend) = test_backend_factory(190, 50);
+
+    let os_input = os_input_output(network_frames, 3);
+    let mut opts = opts_ui();
+    opts.render_opts.total_utilization = true;
     start(backend, os_input, opts);
     let terminal_draw_events_mirror = terminal_draw_events.lock().unwrap();
 
@@ -1097,7 +1318,7 @@ fn layout_full_width_under_30_height() {
             "10.0.0.2",
             1337,
             4435,
-            b"Awesome, I'm from 3.3.3.3",
+            b"Greetings traveller, I'm from 3.3.3.3",
         )),
         Some(build_tcp_packet(
             "2.2.2.2",
@@ -1150,7 +1371,7 @@ fn layout_under_120_width_full_height() {
             "10.0.0.2",
             1337,
             4435,
-            b"Awesome, I'm from 3.3.3.3",
+            b"Greetings traveller, I'm from 3.3.3.3",
         )),
         Some(build_tcp_packet(
             "2.2.2.2",
@@ -1202,7 +1423,7 @@ fn layout_under_120_width_under_30_height() {
             "10.0.0.2",
             1337,
             4435,
-            b"Awesome, I'm from 3.3.3.3",
+            b"Greetings traveller, I'm from 3.3.3.3",
         )),
         Some(build_tcp_packet(
             "2.2.2.2",